fix types.h

+++ types.h
@@ -3,11 +3,11 @@
 
 typedef unsigned char u8;
 typedef unsigned short u16;
-typedef unsigned long u32;
+typedef unsigned int u32;
 typedef unsigned long long u64;
 typedef signed char s8;
 typedef signed short s16;
-typedef signed long s32;
+typedef signed int s32;
 typedef signed long long s64;
 
 typedef float f32;
